refactor(pxi): extract resizable floating panel
